+++ integration/src/main/java/dev/hyperears/integration/EarbudAdapter.kt
@@ -551,6 +551,7 @@ object EarbudAdapterRegistry {
         add(Registration(edifierGroup, ::EdifierHeadphonesAdapter))
         add(Registration(edifierGroup, ::EdifierEarbudAdapter))
         add(Registration(roseGroup, ::FurinaEndlessAdapter))
+        add(Registration(roseGroup, ::RoseLuliUltraAdapter))
         add(Registration(roseGroup, ::RoseEarfreeI5Adapter))
         add(Registration(roseGroup, ::RoseEarfreeProtocolFamilyAdapter))
         add(Registration(roseGroup, ::RoseBudsFeelMk2Adapter))